+++ src/SFA.DAS.Commitments.Notification.WebJob/SFA.DAS.Commitments.Notification.WebJob.csproj
@@ -157,7 +157,7 @@
     <Compile Include="EmailTemplatesService.cs" />
     <Compile Include="IEmailTemplatesService.cs" />
     <Compile Include="INotificationJob.cs" />
-    <Compile Include="INotificationSummary.cs" />
+    <Compile Include="NotificationJob.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Models\UserModel.cs" />
